+++ R/clean_names.R
@@ -72,11 +72,11 @@ clean_names.character <- function(x, include_names = FALSE, ...) {
   if (is.null(x)) {
     return(x)
   }
-  out <- sapply(x, function(.x) {
+  out <- unlist(lapply(x, function(.x) {
     # in case we have ranges, like [2:5], remove those first, so it's not
     # treated as "interaction"
     .x <- sub("\\[(\\d+):(\\d+)\\]", "", .x)
-    if (grepl(":", .x, fixed = TRUE) && !grepl("::", .x, fixed = TRUE)) {
+    if (grepl(":", .x, fixed = TRUE) && !grepl("::", .x, fixed = TRUE) && !startsWith(.x, "mm(")) {
       paste(
         sapply(
           strsplit(.x, ":", fixed = TRUE),
@@ -88,10 +88,10 @@ clean_names.character <- function(x, include_names = FALSE, ...) {
     } else {
       .remove_pattern_from_names(.x, is_emmeans = is_emmeans)
     }
-  })
+  }), use.names = FALSE)
 
   if (isTRUE(include_names)) {
-    out
+    stats::setNames(out, x)
   } else {
     unname(out)
   }
@@ -120,8 +120,8 @@ clean_names.character <- function(x, include_names = FALSE, ...) {
     "logical", "ordered", "as.ordered", "pspline", "scale(poly", "poly", "catg",
     "asis", "matrx", "pol", "strata", "strat", "scale", "scored", "interaction",
     "sqrt", "sin", "cos", "tan", "acos", "asin", "atan", "atan2", "exp", "lsp",
-    "rcs", "pb", "lo", "bs", "ns", "mSpline", "bSpline", "t2", "te", "ti", "tt", # need to be fixed first "mmc", "mm",
-    "mi", "mo", "gp", "s", "I", "gr", "relevel(as.factor", "relevel"
+    "rcs", "pb", "lo", "bs", "ns", "mSpline", "bSpline", "t2", "te", "ti", "tt",
+    "mmc", "mm", "mi", "mo", "gp", "s", "I", "gr", "relevel(as.factor", "relevel"
   )
 
   # sometimes needed for panelr models, where we need to preserve "lag()"
@@ -132,11 +132,14 @@ clean_names.character <- function(x, include_names = FALSE, ...) {
 
   # do we have a "log()" pattern here? if yes, get capture region
   # which matches the "cleaned" variable name
-  cleaned <- sapply(seq_along(x), function(i) {
+  cleaned <- unlist(lapply(seq_along(x), function(i) {
     # check if we have special patterns like 100 * log(xy), and remove it
     if (isFALSE(is_emmeans) && grepl("^([0-9]+)", x[i])) {
       x[i] <- gsub("^([0-9]+)[^(\\.|[:alnum:])]+(.*)", "\\2", x[i])
     }
+    # for brms multimembership, multiple elements might be returned
+    # need extra handling
+    multimembership <- NULL
     for (j in seq_along(pattern)) {
       # check if we find pattern at all
       if (grepl(pattern[j], x[i], fixed = TRUE)) {
@@ -162,10 +165,26 @@ clean_names.character <- function(x, include_names = FALSE, ...) {
         } else if (pattern[j] == "scale(poly") {
           x[i] <- trim_ws(unique(sub("^scale\\(poly\\(((\\w|\\.)*).*", "\\1", x[i])))
         } else if (pattern[j] %in% c("mmc", "mm")) {
-          ## FIXME multimembership-models need to be fixed
-          p <- paste0("^", pattern[j], "\\((.*)\\).*")
-          g <- trim_ws(sub(p, "\\1", x[i]))
-          x[i] <- trim_ws(unlist(strsplit(g, ",", fixed = TRUE), use.names = FALSE))
+          # # detect mm-pattern
+          # p <- paste0("^", pattern[j], "\\((.*)\\).*")
+          # # extract terms from mm() / mmc() functions
+          # g <- trim_ws(sub(p, "\\1", x[i]))
+          # # split terms, but not if comma inside parentheses
+          # g <- trim_ws(unlist(strsplit(g, ",(?![^()]*\\))", perl = TRUE), use.names = FALSE))
+          # # we might have additional arguments, like scale or weights. handle these here
+          # g <- g[!startsWith(g, "scale")]
+          # # clean weights
+          # gweights <- g[startsWith(g, "weights")]
+          # if (length(gweights)) {
+          #   g <- g[!startsWith(g, "weights")]
+          #   # this regular pattern finds "weights=" or "weights =", possibly followed
+          #   # by "cbind()", e.g. "weights = cbind(w, w)". We extract the variable names,
+          #   # create a formula, so "all.vars()" will only extract variable names if
+          #   # we really have "cbind()" in the weights argument
+          #   g <- c(g, .safe(all.vars(as.formula(paste0("~", trim_ws(gsub("weights\\s?=(.*)", "\\1", "weights = cbind(w, w)"))))))) # nolint
+          # }
+          # multimembership <- as.vector(trim_ws(g))
+          multimembership <- all.vars(stats::as.formula(paste("~", x[i])))
         } else if (pattern[j] == "s" && startsWith(x[i], "s(")) {
           x[i] <- gsub("^s\\(", "", x[i])
           x[i] <- gsub("\\)$", "", x[i])
@@ -191,8 +210,13 @@ clean_names.character <- function(x, include_names = FALSE, ...) {
     if (grepl("|", x[i], fixed = TRUE)) {
       x[i] <- sub("^(.*)\\|(.*)", "\\2", x[i])
     }
-    trim_ws(x[i])
-  })
+    # either return regular term, or mm term for brms
+    if (is.null(multimembership)) {
+      trim_ws(x[i])
+    } else {
+      multimembership
+    }
+  }), use.names = FALSE)
 
   # remove for random intercept only models
   .remove_values(cleaned, c("1", "0"))
@@ -210,11 +234,9 @@ clean_names.character <- function(x, include_names = FALSE, ...) {
 
   # extract terms from mm() / mmc() functions, i.e. get
   # multimembership-terms
-  compact_character(unlist(sapply(c("mmc", "mm"), function(j) {
+  compact_character(unlist(lapply(c("mmc", "mm"), function(j) {
     if (grepl(paste0("^", j, "\\("), x = x)) {
-      p <- paste0("^", j, "\\((.*)\\).*")
-      g <- trim_ws(sub(p, "\\1", x))
-      trim_ws(unlist(strsplit(g, ",", fixed = TRUE), use.names = FALSE))
+      all.vars(stats::as.formula(paste("~", x)))
     } else {
       ""
     }